Add support for excel serializers for Pandas and Polars
