+++ .github/workflows/release.yaml
@@ -16,4 +16,4 @@ jobs:
         uses: SneaksAndData/github-actions/semver_release@v0.1.9
         with:
           major_v: 3
-          minor_v: 3
+          minor_v: 4
