+++ adapta/storage/blob/README.md
@@ -8,12 +8,13 @@ In order to init a storage client, you need a respective authentication provider
 
 ### Azure examples
 #### Read multiple blobs into a pandas Dataframe:
+
 ```python
 import pandas
 from adapta.security.clients import AzureClient
 from adapta.storage.models.azure import AdlsGen2Path
 from adapta.storage.blob.azure_storage_client import AzureStorageClient
-from adapta.storage.models.format import DataFrameParquetSerializationFormat
+from adapta.storage.models.formatters import DataFrameParquetSerializationFormat
 
 azure_client = AzureClient()
 adls_path = AdlsGen2Path.from_hdfs_path('abfss://container@account.dfs.core.windows.net/path/to/my/table')
@@ -52,24 +53,25 @@ cat /local/path/file_name
 ```
 
 ### AWS example
+
 ```python
 import os
 from adapta.security.clients import AwsClient
 from adapta.security.clients.aws._aws_credentials import EnvironmentAwsCredentials
 from adapta.storage.models.aws import S3Path
 from adapta.storage.blob.s3_storage_client import S3StorageClient
-from adapta.storage.models.format import DictJsonSerializationFormat
+from adapta.storage.models.formatters import DictJsonSerializationFormat
 
 # Set up environment variables
-os.environ["PROTEUS__AWS_ACCESS_KEY_ID"] = <aws_access_key_id>
-os.environ["PROTEUS__AWS_SESSION_TOKEN"] = <aws_session_token>
-os.environ["PROTEUS__AWS_SECRET_ACCESS_KEY"] = <aws_secret_access_key>
+os.environ["PROTEUS__AWS_ACCESS_KEY_ID"] = < aws_access_key_id >
+os.environ["PROTEUS__AWS_SESSION_TOKEN"] = < aws_session_token >
+os.environ["PROTEUS__AWS_SECRET_ACCESS_KEY"] = < aws_secret_access_key >
 os.environ["PROTEUS__AWS_REGION"] = "eu-central-1"
 os.environ["PROTEUS__AWS_ENDPOINT"] = "http://example.com"
 
 # Create client
 credentials = EnvironmentAwsCredentials()  # Create AWS credentials
-aws_client = AwsClient(credentials) 
+aws_client = AwsClient(credentials)
 
 # Initialize storage client
 s3_client = S3StorageClient.create(auth=aws_client)
@@ -80,7 +82,8 @@ s3_path = S3Path.from_hdfs_path('s3a://bucket/folder/path_to_my_blob')
 # Save data to S3 path
 data = {
     'Character': ['Homer Simpson', 'Michael Scott', 'Ron Swanson', 'Sheldon Cooper', 'Captain Jack Sparrow'],
-    'Occupation': ['Nuclear Safety Inspector', 'Regional Manager', 'Parks and Recreation Director', 'Theoretical Physicist', 'Pirate Captain'],
+    'Occupation': ['Nuclear Safety Inspector', 'Regional Manager', 'Parks and Recreation Director',
+                   'Theoretical Physicist', 'Pirate Captain'],
     'Catchphrase': [
         'D\'oh!',
         'I am the World\'s Best Boss.',
@@ -112,18 +115,19 @@ print(list(folder_data_path_iterator))
 ```
 
 #### Download a single blob:
+
 ```python
 import os
 from adapta.security.clients import AwsClient
 from adapta.security.clients.aws._aws_credentials import EnvironmentAwsCredentials
 from adapta.storage.models.aws import S3Path
 from adapta.storage.blob.s3_storage_client import S3StorageClient
-from adapta.storage.models.format import DataFrameParquetSerializationFormat
+from adapta.storage.models.formatters import DataFrameParquetSerializationFormat
 
 # Set up environment variables
-os.environ["PROTEUS__AWS_ACCESS_KEY_ID"] = <aws_access_key_id>
-os.environ["PROTEUS__AWS_SESSION_TOKEN"] = <aws_session_token>
-os.environ["PROTEUS__AWS_SECRET_ACCESS_KEY"] = <aws_secret_access_key>
+os.environ["PROTEUS__AWS_ACCESS_KEY_ID"] = < aws_access_key_id >
+os.environ["PROTEUS__AWS_SESSION_TOKEN"] = < aws_session_token >
+os.environ["PROTEUS__AWS_SECRET_ACCESS_KEY"] = < aws_secret_access_key >
 os.environ["PROTEUS__AWS_REGION"] = "eu-central-1"
 os.environ["PROTEUS__AWS_ENDPOINT"] = "http://example.com"
