+++ tests/test_serialization_format.py
@@ -18,23 +18,40 @@ from typing import Type
 import pandas
 import polars
 from adapta.storage.models.format import (
-    DictJsonSerializationFormat,
     SerializationFormat,
-    UnitSerializationFormat,
-    PickleSerializationFormat,
-    PolarsDataFrameJsonSerializationFormat,
-    PolarsDataFrameCsvSerializationFormat,
-    PolarsDataFrameParquetSerializationFormat,
-    PandasDataFrameParquetSerializationFormat,
-    PandasDataFrameJsonSerializationFormat,
-    PandasDataFrameCsvSerializationFormat,
-    MetaFrameParquetSerializationFormat,
-    PolarsLazyFrameJsonSerializationFormat,
-    PolarsLazyFrameParquetSerializationFormat,
-    PolarsLazyFrameCsvSerializationFormat,
-    PolarsDataFrameExcelSerializationFormat,
-    PandasDataFrameExcelSerializationFormat,
 )
+from adapta.storage.models.formatters import (
+    DictJsonSerializationFormat,
+    DictJsonSerializationFormatWithFileFormat,
+    MetaFrameParquetSerializationFormat,
+    MetaFrameParquetSerializationFormatWithFileFormat,
+    PandasDataFrameJsonSerializationFormat,
+    PandasDataFrameJsonSerializationFormatWithFileFormat,
+    PandasDataFrameCsvSerializationFormat,
+    PandasDataFrameCsvSerializationFormatWithFileFormat,
+    PandasDataFrameParquetSerializationFormat,
+    PandasDataFrameParquetSerializationFormatWithFileFormat,
+    PandasDataFrameExcelSerializationFormat,
+    PandasDataFrameExcelSerializationFormatWithFileFormat,
+    PickleSerializationFormat,
+    PickleSerializationFormatWithFileFormat,
+    PolarsDataFrameExcelSerializationFormat,
+    PolarsDataFrameExcelSerializationFormatWithFileFormat,
+    PolarsLazyFrameJsonSerializationFormat,
+    PolarsLazyFrameJsonSerializationFormatWithFileFormat,
+    PolarsLazyFrameCsvSerializationFormat,
+    PolarsLazyFrameCsvSerializationFormatWithFileFormat,
+    PolarsLazyFrameParquetSerializationFormat,
+    PolarsLazyFrameParquetSerializationFormatWithFileFormat,
+    PolarsDataFrameJsonSerializationFormat,
+    PolarsDataFrameJsonSerializationFormatWithFileFormat,
+    PolarsDataFrameCsvSerializationFormat,
+    PolarsDataFrameCsvSerializationFormatWithFileFormat,
+    PolarsDataFrameParquetSerializationFormat,
+    PolarsDataFrameParquetSerializationFormatWithFileFormat,
+    UnitSerializationFormat,
+)
+
 from adapta.utils.metaframe import MetaFrame
 
 
@@ -42,46 +59,68 @@ from adapta.utils.metaframe import MetaFrame
     "serializer, data",
     [
         (DictJsonSerializationFormat, {"test": "test"}),
+        (DictJsonSerializationFormatWithFileFormat, {"test": "test"}),
         (
             PandasDataFrameParquetSerializationFormat,
             pandas.DataFrame(data={"test": [1, 2, 3]}),
         ),
         (PandasDataFrameJsonSerializationFormat, pandas.DataFrame(data={"test": [1, 2, 3]})),
+        (PandasDataFrameJsonSerializationFormatWithFileFormat, pandas.DataFrame(data={"test": [1, 2, 3]})),
         (PandasDataFrameCsvSerializationFormat, pandas.DataFrame(data={"test": [1, 2, 3]})),
         (
             PolarsDataFrameParquetSerializationFormat,
             polars.DataFrame(data={"test": [1, 2, 3]}),
         ),
+        (PandasDataFrameCsvSerializationFormatWithFileFormat, pandas.DataFrame(data={"test": [1, 2, 3]})),
         (
             PolarsDataFrameCsvSerializationFormat,
             polars.DataFrame(data={"test": [1, 2, 3]}),
         ),
+        (PolarsDataFrameCsvSerializationFormatWithFileFormat, polars.DataFrame(data={"test": [1, 2, 3]})),
         (
             PolarsDataFrameJsonSerializationFormat,
             polars.DataFrame(data={"test": [1, 2, 3]}),
         ),
+        (PolarsDataFrameJsonSerializationFormatWithFileFormat, polars.DataFrame(data={"test": [1, 2, 3]})),
         (
             PolarsLazyFrameParquetSerializationFormat,
             polars.LazyFrame(data={"test": [1, 2, 3]}),
         ),
+        (PolarsLazyFrameParquetSerializationFormatWithFileFormat, polars.LazyFrame(data={"test": [1, 2, 3]})),
         (
             PolarsLazyFrameCsvSerializationFormat,
             polars.LazyFrame(data={"test": [1, 2, 3]}),
         ),
+        (PolarsLazyFrameCsvSerializationFormatWithFileFormat, polars.LazyFrame(data={"test": [1, 2, 3]})),
         (
             PolarsLazyFrameJsonSerializationFormat,
             polars.LazyFrame(data={"test": [1, 2, 3]}),
         ),
+        (PolarsLazyFrameJsonSerializationFormatWithFileFormat, polars.LazyFrame(data={"test": [1, 2, 3]})),
         (PickleSerializationFormat, pandas.DataFrame(data={"test": [1, 2, 3]})),
+        (PickleSerializationFormatWithFileFormat, pandas.DataFrame(data={"test": [1, 2, 3]})),
         (PickleSerializationFormat, [1, 2, 3]),
+        (PickleSerializationFormatWithFileFormat, [1, 2, 3]),
         (PickleSerializationFormat, {"foo": "bar"}),
+        (PickleSerializationFormatWithFileFormat, {"foo": "bar"}),
         (PickleSerializationFormat, "Hello, World!"),
+        (PickleSerializationFormatWithFileFormat, "Hello, World!"),
         (PickleSerializationFormat, b"Test string"),
-        (UnitSerializationFormat, b"Test string"),
+        (PickleSerializationFormatWithFileFormat, b"Test string"),
         (MetaFrameParquetSerializationFormat, MetaFrame.from_pandas(pandas.DataFrame(data={"test": [1, 2, 3]}))),
+        (
+            MetaFrameParquetSerializationFormatWithFileFormat,
+            MetaFrame.from_pandas(pandas.DataFrame(data={"test": [1, 2, 3]})),
+        ),
         (MetaFrameParquetSerializationFormat, MetaFrame.from_polars(polars.DataFrame(data={"test": [1, 2, 3]}))),
+        (
+            MetaFrameParquetSerializationFormatWithFileFormat,
+            MetaFrame.from_polars(polars.DataFrame(data={"test": [1, 2, 3]})),
+        ),
         (PolarsDataFrameExcelSerializationFormat, polars.DataFrame(data={"test": [1, 2, 3]})),
+        (PolarsDataFrameExcelSerializationFormatWithFileFormat, polars.DataFrame(data={"test": [1, 2, 3]})),
         (PandasDataFrameExcelSerializationFormat, pandas.DataFrame(data={"test": [1, 2, 3]})),
+        (PandasDataFrameExcelSerializationFormatWithFileFormat, pandas.DataFrame(data={"test": [1, 2, 3]})),
     ],
 )
 def test_unit_serialization(serializer: Type[SerializationFormat], data: any):
